Test moving the icons to public
